pyproject.toml
[tool.poetry]
name = "anime-list"
version = "0.1.0"
description = ""
authors = ["Vitória Emanuely"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.8"
gunicorn = "^20.1.0"
whitenoise = "^5.3.0"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^21.10b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
